backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.0

# Database
psycopg2-binary>=2.9.9

# Environment
python-dotenv>=1.0.0

# Utils
Pillow>=10.0.0

# Development
django-debug-toolbar>=4.2.0
